Move GCP Code into module
